Tracking down agg stats failures
