+++ crates/fluvio/src/stats/monitor/mod.rs
@@ -1,2 +1,4 @@
 mod process;
 mod stats;
+mod event;
+pub use event::ClientStatsEvent;
